saveProduto with method Post (OK)

+++ child-play/child-play/src/main/java/br/com/ecommerce/childplay/dao/ProdutoDAO.java
@@ -42,7 +42,7 @@ public class ProdutoDAO {
                 produto.setCaracteristicas(rs.getString("caracteristicas"));
                 produto.setCategoria(rs.getString("categoria"));
                 produto.setIdade(rs.getInt("idade"));
-                produto.setPreco(rs.getFloat("preco"));
+                produto.setPreco(rs.getDouble("preco"));
                 produto.setEstoque(rs.getInt("estoque"));
                 produto.setDesconto(rs.getInt("desconto"));
 
@@ -99,7 +99,7 @@ public class ProdutoDAO {
                 produto.setCaracteristicas(rs.getString("caracteristicas"));
                 produto.setCategoria(rs.getString("categoria"));
                 produto.setIdade(rs.getInt("idade"));
-                produto.setPreco(rs.getFloat("preco"));
+                produto.setPreco(rs.getDouble("preco"));
                 produto.setEstoque(rs.getInt("estoque"));
                 produto.setDesconto(rs.getInt("desconto"));
