+++ child-play/child-play/src/main/java/br/com/ecommerce/childplay/model/ItemPedido.java
@@ -1,13 +1,12 @@
 
 package br.com.ecommerce.childplay.model;
 
-import java.math.BigDecimal;
 
 public class ItemPedido {
     private int idPedido;
     private int idProduto;
     private int quantidade;
-    private BigDecimal preco;
+    private double preco;
 
     public int getIdPedido() {
         return idPedido;
@@ -33,11 +32,11 @@ public class ItemPedido {
         this.quantidade = quantidade;
     }
 
-    public BigDecimal getPreco() {
+    public double getPreco() {
         return preco;
     }
 
-    public void setPreco(BigDecimal preco) {
+    public void setPreco(double preco) {
         this.preco = preco;
     }
 }
